Move migrations to the store and ensure they follow an order mixing database and data migrations
